refact: Changer le type du column 'price' du 'double' vers 'decimal'

+++ database/migrations/0001_01_01_000004_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price');
+            $table->decimal('price');
             $table->foreignId('departments_id')->constrained('departments')->onDelete('cascade');
             $table->boolean('hasDiscount')->default(false);
             $table->timestamps();
